+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 - Adds an event system for ECS operations like entity creation/removal, component addition/removal and relation target changes (#330, #331, #333, #342, #344, #352, #358, #364, #410)
 - Adds custom event support for the event system (#340, #348)
-- Adds method `World.Shrink` for freeing memory that exceeds current requirements (#323)
+- Adds method `World.Shrink` for freeing memory that exceeds current requirements (#323, #417)
 - World lock and filters are concurrency-safe, allowing for concurrent query execution (#360. #405)
 - Adds methods `ObserverX.New`, `Map.New` and `Resource.New` as shortcuts to avoid repeated parameter listing (#393, #394. #395)
 - Adds method `World.CopyEntity` for entity duplication (#415)